:zap:Feat: dropdown 함수 재사용 위해 분리
